Split useful parts out of dropbox sync code
